address partial comments
